pyproject.toml
[project]
name = "llm-workflow-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ollama>=0.6.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
]
